ensure proper data type for variant
gcc10 is more strict on variant usage

+++ dbus/dbuswrite.cpp
@@ -63,7 +63,8 @@ void DbusWritePercent::write(double value)
     auto mesg =
         writeBus.new_method_call(connectionName.c_str(), path.c_str(),
                                  "org.freedesktop.DBus.Properties", "Set");
-    mesg.append(pwmInterface, "Target", std::variant<uint64_t>(ovalue));
+    mesg.append(pwmInterface, "Target",
+                std::variant<uint64_t>(static_cast<uint64_t>(ovalue)));
 
     try
     {
@@ -109,7 +110,8 @@ void DbusWrite::write(double value)
     auto mesg =
         writeBus.new_method_call(connectionName.c_str(), path.c_str(),
                                  "org.freedesktop.DBus.Properties", "Set");
-    mesg.append(pwmInterface, "Target", std::variant<uint64_t>(value));
+    mesg.append(pwmInterface, "Target",
+                std::variant<uint64_t>(static_cast<uint64_t>(value)));
 
     try
     {
